:seedling: divide URLResolver to each class for readability & place GenerateDynamicProjectContextListJson path
