+++ Assets/Scripts/UnityModule/AssetBundleManagement/URLResolver/AmazonCloudFrontURLResolver.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 61dc4b8990c747fea0c849f1ccfdc2d0
+timeCreated: 1602653183
