+++ Assets/Scripts/UnityModule/AssetBundleManagement/URLResolver/AmazonS3URLResolver.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f75bc2b546354a719ab848c033c86aa6
+timeCreated: 1602653140
