+++ Assets/Scripts/UnityModule/AssetBundleManagement/URLResolver/EditorURLResolver.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c03a662292a74327a567810786c90449
+timeCreated: 1602653097
