+++ Assets/Scripts/UnityModule/AssetBundleManagement/URLResolver/IURLResolver.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 0da803cf9c474694a6b4580e6113f9f1
+guid: 560de0a24c0414cb682d51c77dd10f65
 timeCreated: 1519820661
