+++ Assets/Scripts/UnityModule/AssetBundleManagement/URLResolver/URLResolverBase.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a7482874cf81457489848e43b2bab273
+timeCreated: 1602653049
